Created login component and started login fetching user auth
